Previous code summary: scored one in auto.... second one didnt shoot 3 point were more consistant but it appears that the fire button was being pressed before the wheel was truely at speed..
Modifications:

added a second PID shoot command just for autonomous to wait before the
ball is shot... to give it more time to score in auto...

created a custom competition dashboard that shows when the shooter is at
speed.

we noticing during testing that the true time for the wheel to be at
speed is ~5 seconds to guarantee.

this is code ran during 3rd match 13:46 3/17